работает под Ubuntu 13.04 qt5.0.2

+++ src/imagepacker.cpp
@@ -62,7 +62,7 @@ void ImagePacker::pack(int heur, int w, int h)
             summArea -= areaBuf;
         }
         repeat = false;
-        while(summArea > W * H && W * H < w * h)
+        while(summArea > size_t(W * H) && W * H < w * h)
         {
             WH = !WH;
             if(WH)
